# swaggerSpec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WinnerTakesItAll",
    "version": "1.0.0",
    "description": "API documentation of the WinnerTakesItAll project"
  },
  "servers": [
    {
      "url": "http://localhost:3050",
      "description": "Local server"
    }
  ],
  "paths": {
    "/satellites": {
      "post": {
        "summary": "Post satellite data",
        "description": "Posts satellite data to the database. Requires authentication.",
        "tags": [
          "Satellite"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Cookie"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostSatellite"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Satellite data posted successfully"
          },
          "default": {
            "$ref": "#/components/responses/DefaultErrorResponse"
          }
        }
      },
      "get": {
        "summary": "Get satellite data",
        "description": "Gets the data of all satellites as an array for the authenticated user.",
        "tags": [
          "Satellite"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Cookie"
          }
        ],
        "responses": {
          "200": {
            "description": "Satellite data retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetSatellites"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/DefaultErrorResponse"
          }
        }
      }
    },
    "/users/register": {
      "post": {
        "summary": "Register a new user",
        "description": "Creates a new user in the database after validating and hashing the password.",
        "tags": [
          "Users"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegister"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "User registered successfully"
          },
          "default": {
            "$ref": "#/components/responses/DefaultErrorResponse"
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "summary": "User login",
        "description": "Logs in a user with a username and password. Returns a session cookie in the response header.",
        "tags": [
          "Users"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLogin"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/UserLoginResponse"
          },
          "default": {
            "$ref": "#/components/responses/DefaultErrorResponse"
          }
        }
      }
    },
    "/users/logout": {
      "post": {
        "summary": "User logout",
        "description": "Logs out the user by destroying the session.",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Cookie"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/UserLogoutResponse"
          },
          "default": {
            "$ref": "#/components/responses/DefaultErrorResponse"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "PostSatellite": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "satelliteName"
        ],
        "properties": {
          "satelliteName": {
            "type": "string"
          }
        }
      },
      "GetSatellites": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/PostSatellite"
        }
      },
      "UserRegister": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "format": "password"
          }
        }
      },
      "UserLogin": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "format": "password"
          }
        }
      },
      "Message": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      }
    },
    "responses": {
      "DefaultErrorResponse": {
        "description": "Default error response",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "additionalProperties": false,
              "required": [
                "message"
              ],
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "UserLoginResponse": {
        "description": "Login successful",
        "headers": {
          "Set-Cookie": {
            "schema": {
              "type": "string",
              "example": "connect.sid=s%3Ak3f; Path=/; HttpOnly"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Message"
            }
          }
        }
      },
      "UserLogoutResponse": {
        "description": "Logout successful",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Message"
            }
          }
        }
      }
    },
    "parameters": {
      "Cookie": {
        "in": "header",
        "name": "Cookie",
        "schema": {
          "type": "string",
          "example": "connect.sid=s%3Ak3f; Path=/; HttpOnly"
        },
        "description": "Session cookie"
      }
    }
  },
  "tags": []
}